tools/devenv/requirements.txt
dearpygui>=1.11
websocket-client>=1.6
